Separating out focus on first form element functionality into a hook
